moved common dialers to separate file
